requirements.txt
selenium
yeelight
msal
webdriver-manager-4.0.2
